Added filtered view when tapping on pie chart & refactor to use base table view class
